+++ src/molecules/Carousel/Item.tsx
@@ -8,7 +8,8 @@ const propTypes = {
     PropTypes.number
   ]).isRequired,
   children: PropTypes.node.isRequired,
-  onPress: PropTypes.func
+  onPress: PropTypes.func,
+  containerStyles: PropTypes.object
 }
 
 type Props = InferProps<typeof propTypes>
@@ -17,12 +18,14 @@ const defaultProps: Props = {
   onPress: () => {}
 }
 
-function Item ({ id, children, onPress: rawOnPress }: Props) {
+function Item (props: Props) {
+  const { id, children, containerStyles, onPress: rawOnPress } = props
+
   const onPress = useCallback(() => rawOnPress(id), [rawOnPress, id])
 
   return (
     <TouchableWithoutFeedback onPress={onPress}>
-      <View style={styles.container}>
+      <View style={{ ...styles.container, ...containerStyles }}>
         {children}
       </View>
     </TouchableWithoutFeedback>